+++ back/app/Http/Controllers/ParqueoController.php
@@ -10,7 +10,11 @@ class ParqueoController extends Controller
     public function index()
     {
         $parqueos = Parqueo::all();
-        return $parqueos;
+        if( $parqueos->isEmpty()){
+            return response()->json(['message' => 'No se encontraron parqueos.'], 404);
+        } else {
+            return response()->json(['data' =>  $parqueos], 200);
+        }
     }
 
     public function store(Request $request)
@@ -26,13 +30,30 @@ class ParqueoController extends Controller
         $parqueo->numero_de_zonas = $validatedData['numero_de_zonas'];
         $parqueo->mapaParqueo = $validatedData['mapaParqueo'];
 
-        $parqueo->save();
+        try {
+            $parqueo->save();
+            return response()->json([
+                'success' => true,
+                'message' => 'Parqueo creado con éxito',
+                'data' => $parqueo->toArray()
+            ]);
+        } catch (\Exception $e) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Error al crear el parqueo',
+                'error' => $e->getMessage()
+            ]);
+        }
     }
 
     public function show($idParqueo)
     {
         $parqueo = Parqueo::find($idParqueo);
-        return $parqueo;
+        if ($parqueo) {
+            return response()->json(['data' =>$parqueo], 200);
+        } else {
+            return response()->json(['error' => 'No se encontró el parqueo'], 404);
+        }
     }
 
     public function update(Request $request, $idParqueo)
@@ -49,13 +70,22 @@ class ParqueoController extends Controller
         $parqueo->numero_de_zonas = $validatedData['numero_de_zonas'];
         $parqueo->mapaParqueo = $validatedData['mapaParqueo'];
 
-        $parqueo->save();
-        return $parqueo;
+        if($parqueo) {
+            $parqueo->save();
+            return response()->json(['message' => 'Parqueo actualizado correctamente.']);
+        } else {
+            return response()->json(['message' => 'No se encontró el parqueo a actualizar.'], 404);
+        }
     }
 
     public function destroy($idParqueo)
     {
-        $parqueo = Parqueo::destroy($idParqueo);
-        return $parqueo;
+        $parqueo = Parqueo::find($idParqueo);
+        if($parqueo) {
+            $parqueo = Parqueo::destroy($idParqueo);
+            return response()->json(['message' => 'Parqueo eliminado correctamente.']);
+        } else {
+            return response()->json(['message' => 'No se encontró el parqueo a eliminar.'], 404);
+        }
     }
 }
